+++ src/routing/index.ts
@@ -20,5 +20,7 @@ export {
   handleLocationChange,
   type Routing404Event,
 } from "./handleLocationChange";
+export { useIsRouteActive } from "./useIsRouteActive";
+export { useRouteArgsIfActive } from "./useRouteArgsIfActive";
 
-export { RoutingContext } from "./providers";
+export { RoutingContext, TestRoutingContext } from "./providers";