+++ src/test-app/tests/spawningChildActor.integration.tsx
@@ -10,7 +10,7 @@ describe("creating a new child actor", () => {
     await cleanup();
     const { getByTestId, getAllByTestId } = render(<App />);
 
-    await delay(5);
+    await delay(50);
     await act(() => userEvent.type(getByTestId("todo-input"), "test{enter}"));
 
     await delay(300);